requirements.txt
adafruit-circuitpython-mlx90640
picamera2
numpy
opencv-python
pillow
pygame
